+++ com.unity.render-pipelines.universal/Editor/UniversalRenderPipelineCameraEditor.cs
@@ -702,6 +702,12 @@ namespace UnityEditor.Rendering.Universal
                 m_AdditionalCameraDataCameraTypeProp.intValue = selCameraType;
                 m_AdditionalCameraDataSO.ApplyModifiedProperties();
                 UpdateCameras();
+
+                // ScriptableRenderContext.SetupCameraProperties still depends on camera target texture
+                // In order for overlay camera not to override base camera target texture we null it here
+                CameraRenderType camType = (CameraRenderType)m_AdditionalCameraDataCameraTypeProp.intValue;
+                if (camType == CameraRenderType.Overlay && settings.targetTexture.objectReferenceValue != null)
+                    settings.targetTexture.objectReferenceValue = null;
             }
         }